remove separate kyo-os-lib module
