--- kyo-os-lib/shared/src/test/scala/kyo/jvmTest.scala
@@ -1,25 +1,0 @@
-package kyo
-
-import kyo.*
-
-class jvmTest extends KyoTest:
-
-    "execute new jvm and throw error" in run {
-        assert {
-            IOs.run(jvm.process(classOf[MainClass.type], "some-arg" :: Nil).map(_.waitFor)).pure == 1
-        }
-    }
-
-    "execute new jvm and end without error" in run {
-        assert {
-            IOs.run(jvm.process(classOf[MainClass.type]).map(_.waitFor)).pure == 0
-        }
-    }
-
-end jvmTest
-
-object MainClass:
-    def main(args: Array[String]) =
-        if args.isEmpty then System.exit(0)
-        else System.exit(1)
-end MainClass
